Move internal `base-checkbox`
