+++ packages/radix-ui-themes/src/components/_internal/base-checkbox.css
