+++ packages/radix-ui-themes/src/components/_internal/base-checkbox.props.ts
@@ -1,7 +1,7 @@
-import { colorPropDef } from '../props/color.prop.js';
-import { highContrastPropDef } from '../props/high-contrast.prop.js';
+import { colorPropDef } from '../../props/color.prop.js';
+import { highContrastPropDef } from '../../props/high-contrast.prop.js';
 
-import type { PropDef } from '../props/prop-def.js';
+import type { PropDef } from '../../props/prop-def.js';
 
 const sizes = ['1', '2', '3'] as const;
 const variants = ['classic', 'surface', 'soft'] as const;
