+++ packages/radix-ui-themes/src/components/checkbox-cards.css
@@ -1,5 +1,5 @@
 @import './_internal/base-card.css';
-@import './base-checkbox.css';
+@import './_internal/base-checkbox.css';
 
 .rt-CheckboxCardsRoot {
   line-height: var(--line-height);
