+++ packages/radix-ui-themes/src/components/checkbox-cards.tsx
@@ -7,7 +7,7 @@ import { createContextScope } from '@radix-ui/react-context';
 import * as CheckboxGroupPrimitive from './checkbox-group.primitive.js';
 import { createCheckboxGroupScope } from './checkbox-group.primitive.js';
 import { checkboxCardsRootPropDefs } from './checkbox-cards.props.js';
-import { baseCheckboxPropDefs } from './base-checkbox.props.js';
+import { baseCheckboxPropDefs } from './_internal/base-checkbox.props.js';
 import { Grid } from './grid.js';
 import { ThickCheckIcon } from './icons.js';
 import { extractProps } from '../helpers/extract-props.js';
