+++ packages/radix-ui-themes/src/components/checkbox-group.props.ts
@@ -1,5 +1,5 @@
 import { asChildPropDef } from '../props/as-child.prop.js';
-import { baseCheckboxPropDefs } from './base-checkbox.props.js';
+import { baseCheckboxPropDefs } from './_internal/base-checkbox.props.js';
 
 const checkboxGroupRootPropDefs = {
   ...asChildPropDef,
